+++ test/BaseTest.ts
@@ -95,27 +95,3 @@ export const customTest = testBase.extend({
     await use(loadPage);
   }
 });
-
-customTest('Test suite creates a virtual window and DOM', () => {
-  expect(typeof window).not.toBe('undefined');
-  expect(window instanceof Window).toBe(true);
-});
-
-customTest('Fetch mock works properly when enabled.', async () => {
-  fetchMock.enableMocks();
-
-  const mockResponse = {
-    glyphs:'test format',
-    sources:{
-      esri:{
-        url:'more test'
-      }
-    }
-  }
-
-  fetchMock.mockOnce(JSON.stringify(mockResponse))
-  const response = await fetch('https://basemapstyles-api.arcgis.com/arcgis/rest/services/styles/v2/styles/arcgis/navigation');
-  const body = await response.json();
-
-  expect(body).toEqual(mockResponse)
-});
