+++ test/HostedLayer.test.ts
@@ -1,9 +1,9 @@
 //@ts-nocheck
 import { describe, expect, vi, beforeAll, beforeEach } from 'vitest';
-import { customTest as test } from './BaseTest.test';
+import { customTest as test } from './BaseTest';
 import { HostedLayer } from '../src/HostedLayer';
 
-test('Cannot be instantiated directly.', () => {
+test.only('Cannot be instantiated directly.', () => {
   expect(() => {
     const abstractLayer = new HostedLayer({});
   }).toThrowError();
